+++ libs/langgraph/tests/test_kv.py
@@ -1,22 +1,28 @@
 import asyncio
-from typing import Any, List
+from typing import Any
 
 from pytest_mock import MockerFixture
 
-from langgraph.kv.base import BaseKV, KeyValueStore
+from langgraph.kv.base import BaseKV
+from langgraph.kv.batch import AsyncBatchedKV
 
 
-async def test_kv_queue(mocker: MockerFixture) -> None:
+async def test_kv_async_batch(mocker: MockerFixture) -> None:
     aget = mocker.stub()
+    alist = mocker.stub()
 
     class MockKV(BaseKV):
         async def aget(
-            self, pairs: List[tuple[str, str]]
+            self, pairs: list[tuple[str, str]]
         ) -> dict[tuple[str, str], dict[str, Any] | None]:
             aget(pairs)
-            return {pair: {0: pair[0], 1: pair[1]} for pair in pairs}
+            return {pair: 1 for pair in pairs}
 
-    store = KeyValueStore(MockKV())
+        async def alist(self, prefixes: list[str]) -> dict[str, dict[str, Any]]:
+            alist(prefixes)
+            return {prefix: {prefix: 1} for prefix in prefixes}
+
+    store = AsyncBatchedKV(MockKV())
 
     # concurrent calls are batched
     results = await asyncio.gather(
@@ -24,9 +30,18 @@ async def test_kv_queue(mocker: MockerFixture) -> None:
         store.aget([("c", "d")]),
     )
     assert results == [
-        {("a", "b"): {0: "a", 1: "b"}},
-        {("c", "d"): {0: "c", 1: "d"}},
+        {("a", "b"): 1},
+        {("c", "d"): 1},
     ]
     assert [c.args for c in aget.call_args_list] == [
         ([("a", "b"), ("c", "d")],),
+    ]
+
+    results = await asyncio.gather(
+        store.alist(["a", "b"]),
+        store.alist(["c", "d"]),
+    )
+    assert results == [{"a": {"a": 1}, "b": {"b": 1}}, {"c": {"c": 1}, "d": {"d": 1}}]
+    assert [c.args for c in alist.call_args_list] == [
+        (["a", "b", "c", "d"],),
     ]
